Day 6/Week_1_hackathon/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
chromadb==0.4.22
python-multipart==0.0.9
beautifulsoup4==4.12.3
langchain==0.1.9
python-dotenv==1.0.7
pydantic==2.6.1
pydantic-settings==2.1.0
langchain-community==0.0.24
langchain-chroma==0.0.5
google-generativeai==0.3.2
langchain-google-genai==0.0.10
